// src/Components/AddAndEditPerson/interfaces.ts
export interface IMaleData {
  chairman: boolean
  secondChairM: boolean
  firstSpeach: boolean
  gems: boolean
  live: boolean
  studyB: boolean
  studyBReader: boolean
  endPray: boolean
}

export interface IFemaleData {
  portnerOnly: boolean
  secondClassOnly: boolean
  notBibleStudy: boolean
}